+++ src/commands/commands.ts
@@ -1,11 +1,9 @@
-import { Octokit } from "@octokit/core";
 import i18next from "i18next";
-import { MetadataCache, Notice, Platform, TFile, Vault } from "obsidian";
+import { Notice, Platform, TFile } from "obsidian";
 
-import GithubPublisher from "../main";
-import { GithubBranch } from "../publish/branch";
-import { deleteFromGithub } from "../publish/delete";
-import {GitHubPublisherSettings, RepoFrontmatter, Repository, UploadedFiles} from "../settings/interface";
+import { GithubBranch } from "../class/branch";
+import { deleteFromGithub } from "../class/delete";
+import { MonoRepoProperties, MultiRepoProperties, Repository, UploadedFiles} from "../settings/interface";
 import { ListChangedFiles } from "../settings/modals/list_changed";
 import {
 	createLink,
@@ -18,31 +16,17 @@ import {checkRepositoryValidityWithRepoFrontmatter} from "../utils/data_validati
 import { ShareStatusBar } from "../utils/status_bar";
 /**
  * Share all marked note (share: true) from Obsidian to GitHub
- * @param {GithubBranch} PublisherManager
- * @param {GitHubPublisherSettings} settings
- * @param {Octokit} octokit
- * @param {HTMLElement} statusBarItems - The status bar items
- * @param {string} branchName - The branch name created by the plugin
- * @param {RepoFrontmatter} repoFrontmatter
- * @param {TFile[]} sharedFiles - The files marked as shared
- * @param {boolean} createGithubBranch - If the branch has to be created
- * @param plugin
- * @param shortRepo
- * @return {Promise<void>}
  */
 export async function shareAllMarkedNotes(
 	PublisherManager: GithubBranch,
-	settings: GitHubPublisherSettings,
-	octokit: Octokit,
 	statusBarItems: HTMLElement,
 	branchName: string,
-	repoFrontmatter: RepoFrontmatter,
+	monoRepo: MonoRepoProperties,
 	sharedFiles: TFile[],
 	createGithubBranch = true,
-	plugin: GithubPublisher,
-	shortRepo: Repository | null
 ) {
 	const statusBar = new ShareStatusBar(statusBarItems, sharedFiles.length);
+	const repoFrontmatter = monoRepo.frontmatter;
 	try {
 		let errorCount = 0;
 		const fileError : string[] = [];
@@ -50,7 +34,7 @@ export async function shareAllMarkedNotes(
 		if (sharedFiles.length > 0) {
 			const publishedFiles = sharedFiles.map((file) => file.name);
 			if (createGithubBranch) {
-				const isValid = checkRepositoryValidityWithRepoFrontmatter(PublisherManager, settings, repoFrontmatter, sharedFiles.length);
+				const isValid = checkRepositoryValidityWithRepoFrontmatter(PublisherManager, repoFrontmatter, sharedFiles.length);
 				if (!isValid) return false;
 				await PublisherManager.newBranch(branchName, repoFrontmatter);
 			}
@@ -62,8 +46,7 @@ export async function shareAllMarkedNotes(
 						file,
 						false,
 						branchName,
-						repoFrontmatter,
-						shortRepo
+						monoRepo
 					) ;
 					if (uploaded) {
 						listStateUploaded.push(...uploaded.uploaded);
@@ -79,20 +62,17 @@ export async function shareAllMarkedNotes(
 			const noticeValue = `${publishedFiles.length - errorCount} notes`;
 			const deleted = await deleteFromGithub(
 				true,
-				settings,
-				octokit,
 				branchName,
 				PublisherManager,
-				repoFrontmatter,
-				shortRepo
+				monoRepo
 			);
-			
+			const settings = PublisherManager.settings;
 			if (
 				settings.upload.metadataExtractorPath.length > 0 &&
 				Platform.isDesktop
 			) {
 				const metadataExtractor = await getSettingsOfMetadataExtractor(
-					app,
+					PublisherManager.plugin.app,
 					settings
 				);
 				if (metadataExtractor) {
@@ -116,7 +96,7 @@ export async function shareAllMarkedNotes(
 				);
 				if (settings.plugin.displayModalRepoEditing) {
 					const listEdited = createListEdited(listStateUploaded, deleted, fileError);
-					new ListChangedFiles(plugin.app, listEdited).open();
+					new ListChangedFiles(PublisherManager.plugin.app, listEdited).open();
 				}
 			} else {
 				new Notice(
@@ -142,30 +122,24 @@ export async function shareAllMarkedNotes(
  */
 export async function purgeNotesRemote(
 	PublisherManager: GithubBranch,
-	settings: GitHubPublisherSettings,
-	octokit: Octokit,
 	branchName: string,
-	repoFrontmatter: RepoFrontmatter,
-	otherRepo: Repository | null
+	monoRepo: MonoRepoProperties,
 ) {
 	try {
 		new Notice(
-			(i18next.t("informations.startingClean", {repo: repoFrontmatter}))
+			(i18next.t("informations.startingClean", {repo: monoRepo.frontmatter}))
 		);
-		const isValid = checkRepositoryValidityWithRepoFrontmatter(PublisherManager, settings, repoFrontmatter);
+		const isValid = checkRepositoryValidityWithRepoFrontmatter(PublisherManager, monoRepo.frontmatter);
 		if (!isValid) return false;
-		await PublisherManager.newBranch(branchName, repoFrontmatter);
+		await PublisherManager.newBranch(branchName, monoRepo.frontmatter);
 		const deleted = await deleteFromGithub(
 			false,
-			settings,
-			octokit,
 			branchName,
 			PublisherManager,
-			repoFrontmatter,
-			otherRepo
+			monoRepo
 		);
-		await PublisherManager.updateRepository(branchName, repoFrontmatter);
-		if (settings.plugin.displayModalRepoEditing) new ListChangedFiles(app, deleted).open();
+		await PublisherManager.updateRepository(branchName, monoRepo.frontmatter);
+		if (PublisherManager.settings.plugin.displayModalRepoEditing) new ListChangedFiles(PublisherManager.plugin.app, deleted).open();
 	} catch (e) {
 		console.error(e);
 	}
@@ -186,25 +160,28 @@ export async function purgeNotesRemote(
 export async function shareOneNote(
 	branchName: string,
 	PublisherManager: GithubBranch,
-	settings: GitHubPublisherSettings,
 	file: TFile,
 	repository: Repository | null = null,
-	metadataCache: MetadataCache,
-	vault: Vault,
 	title?: string,
-) {
+) {		
+	const settings = PublisherManager.settings;
+	const app = PublisherManager.plugin.app;
+	const metadataCache = app.metadataCache;
 	try {
 		const frontmatter = metadataCache.getFileCache(file)?.frontmatter;
 		const repoFrontmatter = getRepoFrontmatter(settings, repository, frontmatter);
-		const isValid = await checkRepositoryValidityWithRepoFrontmatter(PublisherManager, settings, repoFrontmatter);
+		const isValid = await checkRepositoryValidityWithRepoFrontmatter(PublisherManager, repoFrontmatter);
+		const multiRepo: MultiRepoProperties = {
+			frontmatter: repoFrontmatter,
+			repo: repository
+		};
 		if (!isValid) return false;
 		await PublisherManager.newBranch(branchName, repoFrontmatter);
 		const publishSuccess = await PublisherManager.publish(
 			file,
 			true,
 			branchName,
-			repoFrontmatter,
-			repository,
+			multiRepo,
 			[],
 			true
 		);
@@ -238,11 +215,9 @@ export async function shareOneNote(
 				);
 				await createLink(
 					file,
-					repoFrontmatter,
-					metadataCache,
-					vault,
+					multiRepo,
 					settings,
-					repository
+					app
 				);
 				if (settings.plugin.displayModalRepoEditing) {
 					const listEdited = createListEdited(publishSuccess.uploaded, publishSuccess.deleted, publishSuccess.error);
@@ -277,27 +252,20 @@ export async function shareOneNote(
  */
 export async function shareNewNote(
 	PublisherManager: GithubBranch,
-	octokit: Octokit,
 	branchName: string,
-	vault: Vault,
-	plugin: GithubPublisher,
-	repoFrontmatter: RepoFrontmatter,
-	shortRepo: Repository | null
+	monoRepo: MonoRepoProperties,
 ) {
-	const settings = plugin.settings;
+	const plugin = PublisherManager.plugin;
 	new Notice(i18next.t("informations.scanningRepo") );
-	const sharedFilesWithPaths = PublisherManager.getAllFileWithPath(shortRepo);
+	const sharedFilesWithPaths = PublisherManager.getAllFileWithPath(monoRepo.repo);
 	// Get all file in the repo before the creation of the branch
 	const githubSharedNotes = await PublisherManager.getAllFileFromRepo(
-		repoFrontmatter.branch, // we need to take the master branch because the branch to create doesn't exist yet
-		octokit,
-		settings,
-		repoFrontmatter
+		monoRepo.frontmatter.branch, // we need to take the master branch because the branch to create doesn't exist yet
+		monoRepo.frontmatter
 	);
 	const newlySharedNotes = PublisherManager.getNewFiles(
 		sharedFilesWithPaths,
 		githubSharedNotes,
-		vault
 	);
 	if (newlySharedNotes.length > 0) {
 		new Notice(
@@ -306,64 +274,44 @@ export async function shareNewNote(
 		);
 
 		const statusBarElement = plugin.addStatusBarItem();
-		const isValid = checkRepositoryValidityWithRepoFrontmatter(PublisherManager, settings, repoFrontmatter, newlySharedNotes.length);
+		const isValid = checkRepositoryValidityWithRepoFrontmatter(PublisherManager, monoRepo.frontmatter, newlySharedNotes.length);
 		if (!isValid) return false;
-		await PublisherManager.newBranch(branchName, repoFrontmatter);
+		await PublisherManager.newBranch(branchName, monoRepo.frontmatter);
 		await shareAllMarkedNotes(
 			PublisherManager,
-			plugin.settings,
-			octokit,
 			statusBarElement,
 			branchName,
-			repoFrontmatter,
+			monoRepo,
 			newlySharedNotes,
 			false,
-			plugin,
-			shortRepo
 		);
-	} else {
-		new Notice(i18next.t("informations.noNewNote") );
+		return;
 	}
+	new Notice(i18next.t("informations.noNewNote") );
 }
 
 /**
  * Share edited notes : they exist on the repo, BUT the last edited time in Obsidian is after the last upload. Also share new notes.
- * @param {GithubBranch} PublisherManager
- * @param {Octokit} octokit
- * @param {string} branchName - The branch name created by the plugin
- * @param {Vault} vault
- * @param {GithubPublisher} plugin
- * @param {RepoFrontmatter} repoFrontmatter
- * @param {Repository | null} shortRepo - The repository object if needed (if the repo is from OtherRepo and not the default one)
- * @return {Promise<void>}
  */
 export async function shareAllEditedNotes(
 	PublisherManager: GithubBranch,
-	octokit: Octokit,
 	branchName: string,
-	vault: Vault,
-	plugin: GithubPublisher,
-	repoFrontmatter: RepoFrontmatter,
-	shortRepo: Repository | null
+	monoRepo: MonoRepoProperties,
 ) {
-	const settings = plugin.settings;
+	const plugin = PublisherManager.plugin;
 	new Notice(i18next.t("informations.scanningRepo") );
-	const sharedFilesWithPaths = PublisherManager.getAllFileWithPath(shortRepo);
+	const sharedFilesWithPaths = PublisherManager.getAllFileWithPath(monoRepo.repo);
 	const githubSharedNotes = await PublisherManager.getAllFileFromRepo(
-		repoFrontmatter.branch,
-		octokit,
-		settings,
-		repoFrontmatter
+		monoRepo.frontmatter.branch,
+		monoRepo.frontmatter
 	);
 	const newSharedFiles = PublisherManager.getNewFiles(
 		sharedFilesWithPaths,
 		githubSharedNotes,
-		vault
 	);
 	const newlySharedNotes = await PublisherManager.getEditedFiles(
 		sharedFilesWithPaths,
 		githubSharedNotes,
-		vault,
 		newSharedFiles
 	);
 	if (newlySharedNotes.length > 0) {
@@ -373,20 +321,16 @@ export async function shareAllEditedNotes(
 		);
 
 		const statusBarElement = plugin.addStatusBarItem();
-		const isValid = checkRepositoryValidityWithRepoFrontmatter(PublisherManager, settings, repoFrontmatter, newlySharedNotes.length);
+		const isValid = checkRepositoryValidityWithRepoFrontmatter(PublisherManager, monoRepo.frontmatter, newlySharedNotes.length);
 		if (!isValid) return false;
-		await PublisherManager.newBranch(branchName, repoFrontmatter);
+		await PublisherManager.newBranch(branchName, monoRepo.frontmatter);
 		await shareAllMarkedNotes(
 			PublisherManager,
-			settings,
-			octokit,
 			statusBarElement,
 			branchName,
-			repoFrontmatter,
+			monoRepo,
 			newlySharedNotes,
 			false,
-			plugin,
-			shortRepo
 		);
 	} else {
 		new Notice(i18next.t("informations.noNewNote") );
@@ -395,63 +339,45 @@ export async function shareAllEditedNotes(
 
 /**
  * share **only** edited notes : they exist on the repo, but the last edited time is after the last upload.
- * @param {GithubBranch} PublisherManager
- * @param {Octokit} octokit
- * @param {string} branchName - The branch name created by the plugin
- * @param {Vault} vault
- * @param {GithubPublisher} plugin
- * @param {RepoFrontmatter} repoFrontmatter
- * @param shortRepo
- * @return {Promise<void>}
  */
 export async function shareOnlyEdited(
 	PublisherManager: GithubBranch,
-	octokit: Octokit,
 	branchName: string,
-	vault: Vault,
-	plugin: GithubPublisher,
-	repoFrontmatter: RepoFrontmatter,
-	shortRepo: Repository | null
+	monoRepo: MonoRepoProperties,
 ) {
-	const settings = plugin.settings;
+	const shortRepo = monoRepo.repo;
+	const repoFrontmatter = monoRepo.frontmatter;
 	new Notice(i18next.t("informations.scanningRepo") );
 	const sharedFilesWithPaths = PublisherManager.getAllFileWithPath(shortRepo);
 	const githubSharedNotes = await PublisherManager.getAllFileFromRepo(
 		repoFrontmatter.branch,
-		octokit,
-		settings,
 		repoFrontmatter
 	);
 	const newSharedFiles: TFile[] = [];
 	const newlySharedNotes = await PublisherManager.getEditedFiles(
 		sharedFilesWithPaths,
 		githubSharedNotes,
-		vault,
 		newSharedFiles
 	);
 	if (newlySharedNotes.length > 0) {
 		new Notice(
 			(i18next.t("informations.foundNoteToSend", {nbNotes: newlySharedNotes.length}))
 		);
-		const statusBarElement = plugin.addStatusBarItem();
-		const isValid = checkRepositoryValidityWithRepoFrontmatter(PublisherManager, settings, repoFrontmatter, newlySharedNotes.length);
+		const statusBarElement = PublisherManager.plugin.addStatusBarItem();
+		const isValid = checkRepositoryValidityWithRepoFrontmatter(PublisherManager, repoFrontmatter, newlySharedNotes.length);
 		if (!isValid) return false;
 		await PublisherManager.newBranch(branchName, repoFrontmatter);
 		await shareAllMarkedNotes(
 			PublisherManager,
-			settings,
-			octokit,
 			statusBarElement,
 			branchName,
-			repoFrontmatter,
+			monoRepo,
 			newlySharedNotes,
 			false,
-			plugin,
-			shortRepo
 		);
-	} else {
-		new Notice(i18next.t("informations.noNewNote") );
+		return;
 	}
+	new Notice(i18next.t("informations.noNewNote") );
 }
